+++ lib/main.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:team_draw/pages/home.dart';
-import 'package:team_draw/pages/player_lineup_page.dart';
-import 'package:team_draw/themes/green_theme.dart';
+import 'package:team_draw/modules/home.dart';
+import 'package:team_draw/shared/themes/green_theme.dart';
 
 void main() {
   runApp(const MyApp());
@@ -15,7 +14,7 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'Só Pra Rir FC',
       theme: greenTheme,
-      home: PlayerLineupPage(),
+      home: Home(),
     );
   }
 }
