--- lib/models/player_type.dart
@@ -1,8 +1,0 @@
-class PlayerType{
-
-  String type;
-  String? description;
-
-  PlayerType({required this.type, this.description});
-
-}
