--- lib/shared/main_drawer.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class MainDrawer extends StatelessWidget {
-  const MainDrawer({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return const Drawer();
-  }
-}
