--- lib/shared/messages.dart
@@ -1,3 +1,0 @@
-const newMatch = "NOVA PARTIDA";
-const selectPlayers = "Selecione os Jogadores";
-const drawTeam = "SORTEAR EQUIPES";
