+++ lib/shared/sections/history_match/all_matches_widget.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/cupertino.dart';
+import 'package:team_draw/models/match.dart';
+
+import '../../../data/match_data.dart';
+import 'matches_widget.dart';
+
+class AllMatchesWidget extends StatefulWidget {
+  const AllMatchesWidget({Key? key}) : super(key: key);
+
+  @override
+  State<AllMatchesWidget> createState() => _AllMatchesWidgetState();
+}
+
+class _AllMatchesWidgetState extends State<AllMatchesWidget> {
+  @override
+  Widget build(BuildContext context) {
+    List<Match> matches = allMatches;
+    return
+      ListView.builder(
+        itemBuilder: (context, index) {
+          return MatchesWidget(match: matches[index]);
+        },
+        itemCount: matches.length,
+        scrollDirection: Axis.horizontal,
+      );
+  }
+}
